Split up artifacts reducer
Move FormReducer, JSONReducer and JSONParser to
their own file.

Close DCOS-14630
