+++ plugins/services/src/js/reducers/JSONMultiContainerReducers.js
@@ -1,7 +1,7 @@
 import {JSONReducer as constraints} from './serviceForm/MultiContainerConstraints';
 import {JSONReducer as containers} from './serviceForm/Containers';
 import {JSONReducer as env} from './serviceForm/EnvironmentVariables';
-import {JSONReducer as fetch} from './serviceForm/Artifacts';
+import {JSONReducer as artifacts} from './serviceForm/JSONReducers/Artifacts';
 import {JSONReducer as scaling} from './serviceForm/MultiContainerScaling';
 import {JSONReducer as labels} from './serviceForm/Labels';
 import {JSONReducer as volumes} from './serviceForm/MultiContainerVolumes';
@@ -24,7 +24,7 @@ module.exports = {
       }
     };
   },
-  fetch,
+  artifacts,
   volumes,
   networks,
   ipAddress
