+++ plugins/services/src/js/reducers/JSONSingleContainerParser.js
@@ -1,6 +1,6 @@
 import {JSONParser as container} from './serviceForm/Container';
 import {JSONParser as constraints} from './serviceForm/Constraints';
-import {JSONParser as fetch} from './serviceForm/Artifacts';
+import {JSONParser as fetch} from './serviceForm/JSONParser/Artifacts';
 import {JSONParser as environmentVariables} from './serviceForm/EnvironmentVariables';
 import {JSONParser as externalVolumes} from './serviceForm/ExternalVolumes';
 import {JSONParser as healthChecks} from './serviceForm/HealthChecks';
