+++ plugins/services/src/js/reducers/JSONSingleContainerReducers.js
@@ -1,7 +1,7 @@
 import {JSONReducer as constraints} from './serviceForm/Constraints';
 import {JSONReducer as container} from './serviceForm/Container';
 import {JSONReducer as env} from './serviceForm/EnvironmentVariables';
-import {JSONReducer as fetch} from './serviceForm/Artifacts';
+import {JSONReducer as fetch} from './serviceForm/JSONReducers/Artifacts';
 import {JSONReducer as healthChecks} from './serviceForm/HealthChecks';
 import {JSONReducer as labels} from './serviceForm/Labels';
 import {JSONReducer as portDefinitions} from './serviceForm/PortDefinitions';
