+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/Artifacts.js
@@ -2,9 +2,8 @@ import {
   ADD_ITEM,
   REMOVE_ITEM,
   SET
-} from '../../../../../../src/js/constants/TransactionTypes';
-import Transaction from '../../../../../../src/js/structs/Transaction';
-import {isEmpty} from '../../../../../../src/js/utils/ValidatorUtil';
+} from '../../../../../../../src/js/constants/TransactionTypes';
+import {isEmpty} from '../../../../../../../src/js/utils/ValidatorUtil';
 
 function getJson(data) {
   return data
@@ -50,30 +49,5 @@ module.exports = {
     this.fetch = processTransaction(this.fetch, {type, path, value});
 
     return getJson(this.fetch);
-  },
-
-  JSONParser(state) {
-    if (state.fetch == null) {
-      return [];
-    }
-
-    return state.fetch.reduce(function (memo, item, index) {
-      memo.push(new Transaction( ['fetch'], index, ADD_ITEM ));
-      Object.keys(item).forEach(function (key) {
-        memo.push(new Transaction( ['fetch', index, key], item[key], SET));
-      });
-
-      return memo;
-    }, []);
-  },
-
-  FormReducer(state = [], {type, path, value}) {
-    if (path == null) {
-      return state;
-    }
-
-    state = processTransaction(state, {type, path, value});
-
-    return state;
   }
 };
